Fix missing translation
When rendering konnector view, we always try to render the error messsage, even
if it doesn't exists. This causes a translation error (key value is `null`) and
leave an empty node into the DOM. This commit fix this by testing that an error
message is available before trying to render it.