Get date from import data not from fields
When an import is requested, prior to that commit, if the date to do the first
import was set it was located with the field values. Now it's given a the root
level of the data transmitted. It's because now konnectors can have
field values for several accounts. So looking at this date in fields, doesn't
make sense anymore.