Rewrite parts of Konnectors backend to be asynchronous
This was needed to be able to generate Weboob-backed connectors
automatically and integrate them in the current workflow (conversation
with Cozyweboob is event-based).

Now, when running `npm run dev_tools -- --init`, one gets 252 konnectors
out of the box, most of them being based on Weboob modules. At the
moment, most of them will not be effective as necessary converters have
not yet been implemented and `fetchOperations` have not been fully
tested.

At the moment, I did not took care of the unit tests, but some of them
will have to be fixed.

Note: I am not 100% satisfied by this commit. At least, it makes it
work. But I'd rather go with `Promise` objects, which would be cleaner
in my opinion and avoid a lot of recomputation (anytime `konnector_hash`
is invoked somewhere, it is recomputed at the moment), but needs a much
more important refactoring. :/